- Add extra filters. - Add extra display data. - Fix visual bug.
