+++ app/templates/layout/base.html
@@ -15,7 +15,7 @@
             <!-- Toggle button -->
             </button>
             <a href="" class="flex items-center justify-between gap-4 mr-4 transition-all duration-100 transform hover:scale-105">
-              <i class="fa-solid fa-book-bookmark text-3xl text-white mt-1"></i> 
+              <!-- <i class="fa-solid fa-book-bookmark text-3xl text-white mt-1"></i>  -->
               <span class="self-center text-2xl font-semibold whitespace-nowrap text-white">100 años de soledad</span>
             </a>
           </div>
